Corrected support for 6.4.3 & using 'installer'
Corrected an issue for 6.4.3 where 'videos.json' did not exist but is required, also updated so that now the download is for the Git repo, but only the installer is used (resources/installer/)

+++ run.sh
@@ -1,7 +1,18 @@
 #!/bin/bash
 echo "Running run.sh startup script"
 declare -a configFiles=("settings.json" "broken.json" "float.log" "videos.json.backup" "videos.json" "partial.json")
+#declare -a configFiles=("settings.json")
 configDirectory="/Floatplane-Downloader/config"
+
+
+exit_script(){
+	clear
+	echo "copying out config files (if new)"
+	copy_config_files
+}
+
+
+
 
 # Setup the config directory, if not already created. This allows for out-of-docker testing
 if [ ! -d $configDirectory ]; then
@@ -10,16 +21,19 @@ fi
 
 arraylength=${#configFiles[@]}
 # check if each of the above files exist in the correct location, if not link or copy them over
+
+copy_config_files(){
 for (( i=1; i<${arraylength}+1; i++ ));
 do
 	if [ ! -f $configDirectory/${configFiles[$i-1]} ]; then
 		# currently the settings files do not exist, they need to be copied over
 		if [ ! -f "/Floatplane-Downloader/${configFiles[$i-1]}" ]; then
-			echo "Missing critical file '/Floatplane-Downloader/${configFiles[$i-1]}', you'l need this to run, creating it now"
-			#echo "{}" > $configDirectory/${configFiles[$i-1]}
-			ln -s $configDirectory/${configFiles[$i-1]} "/Floatplane-Downloader/${configFiles[$i-1]}"
+			if [ -f $configDirectory/${configFiles[$i-1]} ]; then
+				echo "Missing critical file '/Floatplane-Downloader/${configFiles[$i-1]}', you'l need this to run, creating it now"
+				#echo "{}" > $configDirectory/${configFiles[$i-1]}
+				ln -s $configDirectory/${configFiles[$i-1]} "/Floatplane-Downloader/${configFiles[$i-1]}"
+			fi
 		else
-
 			# file exists, but not in the config directory; then link it back
 			mv "/Floatplane-Downloader/${configFiles[$i-1]}" $configDirectory/${configFiles[$i-1]}
 			ln -s $configDirectory/${configFiles[$i-1]} "/Floatplane-Downloader/${configFiles[$i-1]}"
@@ -34,4 +48,8 @@ do
 
   #echo $i " / " ${arraylength} " : " ${array[$i-1]}
 done
+}
+copy_config_files
+
+trap exit_script 2 15 SIGINT SIGTERM
 node float.js
